Port litebox_shim_linux, litebox_platform_linux_userland, and litebox_shim_optee to aarch64
Adds full aarch64 support: ELF loading, syscall dispatch via seccomp+SIGSYS
fallback, signal frame construction and delivery, page-fault/exception
routing, thread/process init, and the guest/host context-switch trampoline
(naked asm) with TPIDR_EL0-based TLS. Fixes a SIGEV_SIGNAL timer-delivery
race (switched to SIGEV_THREAD_ID) on both x86_64 and aarch64. Live-verified
on aarch64 hardware: real ELF binaries executing syscalls through the
seccomp fallback path and receiving correctly delivered SIGSEGV/SIGILL.

+++ litebox_shim_linux/src/syscalls/file.rs
@@ -1820,7 +1820,7 @@ where
         st_gid: 0,
         st_rdev: 0,
         st_size: 0,
-        st_blksize: blksize,
+        st_blksize: blksize as i32,
         st_blocks: 0,
         ..Default::default()
     };
